refactor(database): rename files
